64-bit cleanliness compilation fixes, fix encoding settings in project.

+++ mac/ChromeOSRecovery/AppController.h
@@ -63,7 +63,7 @@
 - (IBAction)nextTab:(id)sender;
 - (IBAction)previousTab:(id)sender;
 - (IBAction)done:(id)sender;
-- (void)switchToTabAtIndex:(int)index;
+- (void)switchToTabAtIndex:(NSInteger)index;
 
 - (IBAction)showHelp:(id)sender;
